requirements.txt
pillow==11.0.0
PyMuPDF==1.24.13
